Import RubyGems 1.1.0
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@15873 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ lib/rubygems/specification.rb
@@ -13,7 +13,7 @@ require 'rubygems/platform'
 if RUBY_VERSION < '1.9' then
   def Time.today
     t = Time.now
-    t - ((t.to_i + t.gmt_offset) % 86400)
+    t - ((t.to_f + t.gmt_offset) % 86400)
   end unless defined? Time.today
 end
 # :startdoc:
